[ANDROID] Moved overlay view to ‘Full’ flavor
