+++ Native/Android/LunarConsole/app/src/androidTest/java/spacemadness/com/lunarconsoleapp/ApplicationBaseUITest.java
@@ -51,7 +51,6 @@ import java.util.List;
 import spacemadness.com.lunarconsole.console.ConsoleCollapsedLogEntry;
 import spacemadness.com.lunarconsole.console.ConsoleLogEntry;
 import spacemadness.com.lunarconsole.console.ConsolePlugin;
-import spacemadness.com.lunarconsole.console.IdentityEntry;
 import spacemadness.com.lunarconsole.console.VariableType;
 import spacemadness.com.lunarconsole.debug.TestHelper;
 import spacemadness.com.lunarconsole.utils.StringUtils;
@@ -672,107 +671,6 @@ public class ApplicationBaseUITest implements TestHelper.EventListener
                         .check(matches(withText(String.valueOf(messageCount[i]))));
             }
         }
-    }
-
-    protected void assertEntries(String[] actions, var[] variables)
-    {
-        int expectedCount = 0;
-        if (actions.length > 0) expectedCount += 1 + actions.length;
-        if (variables.length > 0) expectedCount += 1 + variables.length;
-
-        ViewInteraction listView = onView(withParent(withId(R.id.lunar_console_action_view_list_container)));
-
-        // should be visible
-        listView.check(matches(isDisplayed()));
-
-        // should contains expected number of children
-        listView.check(matches(withListViewSize(expectedCount)));
-
-        if (actions.length > 0)
-        {
-            // first entry is header
-            DataInteraction headerView = onData(allOf(is(instanceOf(IdentityEntry.class))))
-                    .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
-                    .atPosition(0);
-
-            // check title
-            headerView
-                    .onChildView(withId(R.id.lunar_console_header_entry_name))
-                    .check(matches(withText(R.string.lunar_console_header_actions)));
-
-            for (int i = 0; i < actions.length; ++i)
-            {
-                // find entry view
-                DataInteraction entryView = onData(allOf(is(instanceOf(IdentityEntry.class))))
-                        .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
-                        .atPosition(i + 1);
-
-                // check message
-                entryView
-                        .onChildView(withId(R.id.lunar_console_action_entry_name))
-                        .check(matches(withText(actions[i])));
-            }
-        }
-
-        if (variables.length > 0)
-        {
-            int offset = actions.length > 0 ? 1 + actions.length : 0;
-
-            // first entry is header
-            DataInteraction headerView = onData(allOf(is(instanceOf(IdentityEntry.class))))
-                    .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
-                    .atPosition(offset);
-
-            // check title
-            headerView
-                    .onChildView(withId(R.id.lunar_console_header_entry_name))
-                    .check(matches(withText(R.string.lunar_console_header_variables)));
-
-            for (int i = 0; i < variables.length; ++i)
-            {
-                // find entry view
-                DataInteraction entryView = onData(allOf(is(instanceOf(IdentityEntry.class))))
-                        .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
-                        .atPosition(offset + 1 + i);
-
-                // check name
-                entryView
-                        .onChildView(withId(R.id.lunar_console_variable_entry_name))
-                        .check(matches(withText(variables[i].name)));
-
-                switch (variables[i].type)
-                {
-                    case String:
-                    case Integer:
-                    case Float:
-                        // check value
-                        entryView
-                                .onChildView(withId(R.id.lunar_console_variable_entry_value))
-                                .check(matches(withText(variables[i].value)));
-                        break;
-                    case Boolean:
-                        // check flag
-                        entryView
-                                .onChildView(withId(R.id.lunar_console_variable_entry_switch))
-                                .check(matches(variables[i].value.equals("0") ? isNotChecked() : isChecked()));
-                        break;
-                    default:
-                        throw new AssertionError("Unexpected type: " + variables[i].type);
-                }
-            }
-        }
-    }
-
-    protected void clickAction(int index)
-    {
-        ViewInteraction listView = onView(withParent(withId(R.id.lunar_console_action_view_list_container)));
-
-        // should be visible
-        listView.check(matches(isDisplayed()));
-
-        onData(allOf(is(instanceOf(IdentityEntry.class))))
-                .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
-                .atPosition(1 + index).perform(click());
     }
 
     protected MainActivity getActivity()
