+++ Native/Android/LunarConsole/app/src/androidTestFull/java/spacemadness/com/lunarconsoleapp/ConsoleActionsViewTest.java
@@ -21,6 +21,8 @@
 
 package spacemadness.com.lunarconsoleapp;
 
+import android.support.test.espresso.DataInteraction;
+import android.support.test.espresso.ViewInteraction;
 import android.support.test.runner.AndroidJUnit4;
 import android.test.suitebuilder.annotation.LargeTest;
 
@@ -31,11 +33,16 @@ import java.lang.reflect.Field;
 import java.util.Map;
 
 import spacemadness.com.lunarconsole.console.ConsolePlugin;
+import spacemadness.com.lunarconsole.console.IdentityEntry;
 import spacemadness.com.lunarconsole.core.NotificationCenter;
 import spacemadness.com.lunarconsoleapp.helpers.SyncDispatchQueue;
 
+import static android.support.test.espresso.Espresso.*;
 import static android.support.test.espresso.action.ViewActions.*;
-import static spacemadness.com.lunarconsole.debug.TestHelper.TEST_EVENT_NATIVE_CALLBACK;
+import static android.support.test.espresso.matcher.ViewMatchers.*;
+import static android.support.test.espresso.assertion.ViewAssertions.*;
+import static org.hamcrest.Matchers.*;
+import static spacemadness.com.lunarconsole.debug.TestHelper.*;
 
 @RunWith(AndroidJUnit4.class)
 @LargeTest
@@ -780,6 +787,107 @@ public class ConsoleActionsViewTest extends ApplicationBaseUITest
 
     ////////////////////////////////////////////////////////////////////////////////////////////////
     // Helpers
+
+    private void assertEntries(String[] actions, var[] variables)
+    {
+        int expectedCount = 0;
+        if (actions.length > 0) expectedCount += 1 + actions.length;
+        if (variables.length > 0) expectedCount += 1 + variables.length;
+
+        ViewInteraction listView = onView(withParent(withId(R.id.lunar_console_action_view_list_container)));
+
+        // should be visible
+        listView.check(matches(isDisplayed()));
+
+        // should contains expected number of children
+        listView.check(matches(withListViewSize(expectedCount)));
+
+        if (actions.length > 0)
+        {
+            // first entry is header
+            DataInteraction headerView = onData(allOf(is(instanceOf(IdentityEntry.class))))
+                    .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
+                    .atPosition(0);
+
+            // check title
+            headerView
+                    .onChildView(withId(R.id.lunar_console_header_entry_name))
+                    .check(matches(withText(R.string.lunar_console_header_actions)));
+
+            for (int i = 0; i < actions.length; ++i)
+            {
+                // find entry view
+                DataInteraction entryView = onData(allOf(is(instanceOf(IdentityEntry.class))))
+                        .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
+                        .atPosition(i + 1);
+
+                // check message
+                entryView
+                        .onChildView(withId(R.id.lunar_console_action_entry_name))
+                        .check(matches(withText(actions[i])));
+            }
+        }
+
+        if (variables.length > 0)
+        {
+            int offset = actions.length > 0 ? 1 + actions.length : 0;
+
+            // first entry is header
+            DataInteraction headerView = onData(allOf(is(instanceOf(IdentityEntry.class))))
+                    .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
+                    .atPosition(offset);
+
+            // check title
+            headerView
+                    .onChildView(withId(R.id.lunar_console_header_entry_name))
+                    .check(matches(withText(R.string.lunar_console_header_variables)));
+
+            for (int i = 0; i < variables.length; ++i)
+            {
+                // find entry view
+                DataInteraction entryView = onData(allOf(is(instanceOf(IdentityEntry.class))))
+                        .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
+                        .atPosition(offset + 1 + i);
+
+                // check name
+                entryView
+                        .onChildView(withId(R.id.lunar_console_variable_entry_name))
+                        .check(matches(withText(variables[i].name)));
+
+                switch (variables[i].type)
+                {
+                    case String:
+                    case Integer:
+                    case Float:
+                        // check value
+                        entryView
+                                .onChildView(withId(R.id.lunar_console_variable_entry_value))
+                                .check(matches(withText(variables[i].value)));
+                        break;
+                    case Boolean:
+                        // check flag
+                        entryView
+                                .onChildView(withId(R.id.lunar_console_variable_entry_switch))
+                                .check(matches(variables[i].value.equals("0") ? isNotChecked() : isChecked()));
+                        break;
+                    default:
+                        throw new AssertionError("Unexpected type: " + variables[i].type);
+                }
+            }
+        }
+    }
+
+    private void clickAction(int index)
+    {
+        ViewInteraction listView = onView(withParent(withId(R.id.lunar_console_action_view_list_container)));
+
+        // should be visible
+        listView.check(matches(isDisplayed()));
+
+        onData(allOf(is(instanceOf(IdentityEntry.class))))
+                .inAdapterView(withParent(withId(R.id.lunar_console_action_view_list_container)))
+                .atPosition(1 + index).perform(click());
+    }
 
     private void assertNoActions()
     {
