+++ Native/Android/LunarConsole/app/src/androidTestFull/java/spacemadness/com/lunarconsoleapp/ConsoleOverlayLogViewTest.java
