+++ Native/Android/LunarConsole/lunarConsole/src/full/java/spacemadness/com/lunarconsole/console/ConsoleOverlayLogView.java
