+++ Native/Android/LunarConsole/lunarConsole/src/full/res/layout/lunar_console_layout_overlay_log_entry.xml
